Add parkCodes to separate data file
